src/backend/requirements.txt
PyYAML == 6.0.2
psycopg[binary] == 3.2.9